+++ lib/services/database/database.service.d.ts
@@ -1,8 +1,8 @@
-import { ConnectionOptions } from 'mongoose';
+/// <reference types="mongoose" />
 import { ConfigService } from '../config/config.service';
 export declare class DatabaseService {
     private configService;
     constructor(configService: ConfigService);
     connect(): Promise<import("mongodb").Db>;
-    mongooseConnect(options?: ConnectionOptions): Promise<typeof import("mongoose")>;
+    mongooseConnect(): Promise<typeof import("mongoose")>;
 }