+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 * v0.7.6
 - Added `typescript` support
 - Added `init` script helping us to setup project which doesn't have configuration
-- 
+- Removed Error message `DeprecationWarning: current URL string parser is deprecated` from `Mongoose.connect`
 
 * v0.7.5
 - Init commit